requirements.txt
google-generativeai
faiss-cpu
python-dotenv
PyPDF2
numpy
